backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-pptx==0.6.23
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-multipart==0.0.6
Pillow==10.1.0